requirements.txt
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
numpy==2.3.5
packaging==25.0
pandas==2.3.3
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
